feat: custom progress bar added to show overall active/inactive ratio
