--- app/components/Footer.css
@@ -1,9 +1,0 @@
-.footerWrapper {
-  width: 100vw;
-  height: 3em;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding-top: 0.5em;
-  border-top: 2px solid #3c3c3c;
-}
