+++ app/components/ProcessList.css
@@ -3,7 +3,7 @@
   flex-direction: column;
   align-items: center;
   overflow-y: auto;
-  height: calc(100vh - 10em);
+  height: calc(100vh - 7em);
   padding-top: 0.5em;
 }
